requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0
boto3>=1.29.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-multipart>=0.0.6
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
mypy>=1.7.0
black>=23.12.0
# Hugging Face for real CV segmentation
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
# For image processing
scikit-image>=0.22.0
# Dedalus Labs SDK for AI agents
dedalus-labs
# PDF parsing
pypdf>=3.17.0